% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{add_numbers}
\alias{add_numbers}
\title{add numbers}
\usage{
add_numbers(x, y)
}
\arguments{
\item{x}{first number}

\item{y}{second number}
}
\value{
sum of numbers
}
\description{
add numbers
}
\examples{
temp1 <- add_numbers(22);
temp2 <- C_to_F( c(-2, 12, 23) );
}
