% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{C2F}
\alias{C2F}
\title{Celsius conversion}
\usage{
C2F(cTemp)
}
\arguments{
\item{cTemp}{The temperature in degrees Celsius}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Convert degrees Celsius temperatures to degrees Fahrenheit
}
\examples{
temp1 <- C2F(22)
temp2 <- C2F(c(-2, 12, 23))
}
\seealso{
Other Conversation functions: 
\code{\link{F2C}()}
}
\concept{Conversation functions}
