% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{C_to_F}
\alias{C_to_F}
\title{Celsius conversion}
\usage{
C_to_F(C_temp)
}
\arguments{
\item{C_temp}{The temperature in degrees Celsius}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
This function is deprecated, use \code{\link[=C2F]{C2F()}}.
Convert degrees Celsius temperatures to degrees Fahrenheit
}
\seealso{
Other Deprecated functions: 
\code{\link{F_to_C}()}
}
\concept{Deprecated functions}
