% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{f1}
\alias{f1}
\title{My generic function}
\usage{
f1(arg1, arg2, arg3)
}
\arguments{
\item{arg1}{Description of arg1}

\item{arg2}{Description of arg2}

\item{arg3}{Description of arg3}
}
\description{
My generic function
}
\seealso{
Other Test functions: 
\code{\link{f2}()}
}
\concept{Test functions}
