% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_functions.R
\name{f2}
\alias{f2}
\title{My generic function2}
\usage{
f2(arg2, arg3, arg4)
}
\arguments{
\item{arg2}{Description of arg2}

\item{arg3}{Description of arg3}

\item{arg4}{Description of arg4}
}
\description{
My generic function2
}
\seealso{
Other Test functions: 
\code{\link{f1}()}
}
\concept{Test functions}
