% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{F2C}
\alias{F2C}
\title{Fahrenheit conversion}
\usage{
F2C(fTemp)
}
\arguments{
\item{fTemp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
Convert degrees Fahrenheit temperatures to degrees Celsius
}
\examples{
temp1 <- F2C(50)
temp2 <- F2C(c(50, 63, 23))
}
\seealso{
Other Conversation functions: 
\code{\link{C2F}()}
}
\concept{Conversation functions}
