% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{F_to_C}
\alias{F_to_C}
\title{Fahrenheit conversion}
\usage{
F_to_C(F_temp)
}
\arguments{
\item{F_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
This function is deprecated, use \code{\link[=F2C]{F2C()}}.
Convert degrees Fahrenheit temperatures to degrees Celsius
}
\seealso{
Other Deprecated functions: 
\code{\link{C_to_F}()}
}
\concept{Deprecated functions}
